requirements.txt
Django>=4.1,<4.2
wagtail>=4.1,<4.2
psycopg2-binary>=2.9.5
python-dotenv==0.21.0
